Update dev_pins.md
Bedrockminer has left the scene and as such his tutorials are not longer available in the old way. I updated the link to point to the Archive.org version of the page.

The 1.12.2 Configuration Test was no longer in existence, I updated it with new path and filename in the Forge repository. Ironically it still does not use the ANNOTATION system.