Update dev_pins.md
Didn't add the following pin 

When doing a PR to forge, DONT mention asm at all, not even "This would allow me to stop using ASM".
 LexManos commented an hour ago
```If you would like to readdress this as a series of prebuilt utilities to render extra bars or numbers over items in the inventory. Then we can readdress it as that.
But moving things back 5 years by encouraging people to do even more direct GL access is not going to fly.
Feel free to create a new issue or PR to discuss that avenue of things.

As for the whole topic of ASM. Let me state this directly and for the record.
It is, by definition, possible to do ANYTHING in a coremod.
Stating that you can/are using a coremod in a PR serves literally NO purpose.
It is NOTHING but a threat that gets tranted to "Do this or i'll hack shit with a coremod"
I do not bed to threats.

So, new guideline (I've instructed Mezz to make it official). Any mention of ASM or hints at using a coremod to do shit in your issue will result in the thread being locked, and potentially you being banned from the Forge repo.
End of story, i'm sick of being threatened by people and coremods.
From #4803.```